Average Draft information

ADP or AAV results, including when the result were last updated, how many drafts or auctions the player was selected in and the average auction value or average, maximum and minimum pick position. Average auction value is relative to an auction where a total of $1000 is available across all franchises.

Usage

average_draft(
  type = c("adp", "aav"),
  period = c("ALL", "RECENT", "DRAFT", "JUNE", "JULY", "AUG1", "AUG15", "START", "MID",
    "PLAYOFF"),
  num_franchises = c("12", "8", "10", "14", "16"),
  PPR = c("-1", "0", "1"),
  keeper = c("N", "K", "R"),
  mock = c("-1", "0", "1"),
  cutoff = NULL,
  details = NULL
)

Arguments

type

Return average draft position (adp) or average auction values (aav)

period

This returns draft data for just drafts that started after the specified period. Valid values are ALL, RECENT, DRAFT, JUNE, JULY, AUG1, AUG15, START, MID, PLAYOFF.

num_franchises

This returns draft data from just leagues with this number of franchises. Valid values are 8, 10, 12, 14 or 16. If the value is 8, it returns data from leagues with 8 or less franchises. If the value is 16 it returns data from leagues with 16 or more franchises.

PPR

Filters the data returned as follows: If set to 0, data is from leagues that not use a PPR scoring system; if set to 1, only from PPR scoring system; if set to -1, all leagues.

keeper

Pass a string with some combination of N, K and R: if N specified, returns data from redraft leagues, if 'K' is specified, returns data for keeper leagues and if 'R' is specified, return data from rookie-only drafts. You can combine these. If you specify c('K','R') it will return rookie and keeper drafts only. Default is c('N', 'K', 'R').

mock

If set to 1, returns data from mock draft leagues only. If set to 0, excludes data from mock draft leagues. If set to -1, returns all

cutoff

Only returns data for players selected in at least this percentage of drafts. So if you pass 10, it means that players selected in less than 10 Note that if the value is less than 5, the results may be unpredictable.

details

If set to 1, it returns the leagues that were included in the results.