MFL Endpoint

R6 Class representing the the endpoint for MFL. Used for low level operations such as logging in and generating request urls.

Methods

Public methods

mfl_endpoint$new()

mfl_endpoint$print()

mfl_endpoint$login()

mfl_endpoint$export_query()

mfl_endpoint$import_query()

mfl_endpoint$request()

mfl_endpoint$export_url()

mfl_endpoint$import_url()

mfl_endpoint$logged_in()

mfl_endpoint$clone()

Method new()

Usage

mfl_endpoint$new(
  scheme = "https",
  host = "api.myfantasyleague.com",
  season = "2021"
)

Method print()

Usage

mfl_endpoint$print()

Method login()

Usage

mfl_endpoint$login(user_name, password = NULL)

Method export_query()

Usage

mfl_endpoint$export_query(qry_type, qry_params = list())

Method import_query()

Usage

mfl_endpoint$import_query(qry_type, qry_params = list())

Method request()

Usage

mfl_endpoint$request(request_url)

Method export_url()

Usage

mfl_endpoint$export_url()

Method import_url()

Usage

mfl_endpoint$import_url()

Method logged_in()

Usage

mfl_endpoint$logged_in()

Method clone()

The objects of this class are cloneable with this method.

Usage

mfl_endpoint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.